follow pr suggestions
