+++ avaje-record-builder-core/pom.xml
@@ -26,16 +26,9 @@
       <scope>provided</scope>
     </dependency>
     <dependency>
-      <groupId>io.avaje</groupId>
-      <artifactId>avaje-lang</artifactId>
-      <version>1.1</version>
-      <optional>true</optional>
-      <scope>provided</scope> 
-    </dependency>
-    <dependency>
       <groupId>org.jspecify</groupId>
       <artifactId>jspecify</artifactId>
-      <version>0.3.0</version>
+      <version>1.0.0</version>
       <optional>true</optional>
       <scope>provided</scope> 
     </dependency>
