+++ avaje-record-builder-core/src/main/java/io/avaje/recordbuilder/internal/RecordProcessor.java
@@ -15,7 +15,6 @@ import static java.util.stream.Collectors.toMap;
 import java.io.IOException;
 import java.io.UncheckedIOException;
 import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 
 import javax.annotation.processing.AbstractProcessor;
@@ -37,7 +36,7 @@ import io.avaje.prism.GenerateUtils;
 @SupportedAnnotationTypes({
   RecordBuilderPrism.PRISM_TYPE,
   ImportPrism.PRISM_TYPE,
-  GlobalConfigPrism.PRISM_TYPE
+  GlobalPrism.PRISM_TYPE
 })
 public final class RecordProcessor extends AbstractProcessor {
 
@@ -50,31 +49,12 @@ public final class RecordProcessor extends AbstractProcessor {
   public synchronized void init(ProcessingEnvironment env) {
     super.init(env);
     APContext.init(env);
-    GlobalSettings.init();
   }
 
   @Override
   public boolean process(Set<? extends TypeElement> tes, RoundEnvironment roundEnv) {
 
     APContext.setProjectModuleElement(tes, roundEnv);
-    if (!GlobalSettings.initialized()) {
-
-      roundEnv.getElementsAnnotatedWith(typeElement(GlobalConfigPrism.PRISM_TYPE)).stream()
-          .map(GlobalConfigPrism::getInstanceOn)
-          .findFirst()
-          .or(
-              () -> {
-                var module = APContext.getProjectModuleElement();
-                return GlobalConfigPrism.getOptionalOn(module)
-                    .or(
-                        () ->
-                            module.getEnclosedElements().stream()
-                                .map(GlobalConfigPrism::getOptionalOn)
-                                .findAny()
-                                .orElse(Optional.empty()));
-              })
-          .ifPresent(GlobalSettings::configure);
-    }
     final var globalTypeInitializers =
         roundEnv.getElementsAnnotatedWith(typeElement(GlobalPrism.PRISM_TYPE)).stream()
             .map(GlobalPrism::getInstanceOn)
@@ -108,7 +88,6 @@ public final class RecordProcessor extends AbstractProcessor {
       } catch (IOException e) {
         // Can't read module, it's whatever
       }
-      GlobalSettings.clear();
       APContext.clear();
     }
     return false;
@@ -141,7 +120,7 @@ public final class RecordProcessor extends AbstractProcessor {
               .collect(joining(", "))
               .transform(s -> s.isEmpty() ? s : "<" + s + ">");
       writer.append(
-          ClassBodyBuilder.createClassStart(prism, type, typeParams, isImported, packageName));
+          ClassBodyBuilder.createClassStart(type, typeParams, isImported, packageName));
 
       methods(writer, typeParams, shortName, components, prism);
     } catch (final IOException e) {
@@ -155,7 +134,7 @@ public final class RecordProcessor extends AbstractProcessor {
       String shortName,
       List<? extends RecordComponentElement> components,
       BuilderPrism prism) {
-    boolean getters = GlobalSettings.getters() || prism.getters();
+    boolean getters = prism.getters();
 
     for (final var element : components) {
       final var type = UType.parse(element.asType());
