+++ avaje-record-builder-core/src/main/java/io/avaje/recordbuilder/internal/Utils.java
@@ -1,13 +1,5 @@
 package io.avaje.recordbuilder.internal;
 
-import static java.util.Map.entry;
-
-import java.util.Map;
-import java.util.Optional;
-import java.util.OptionalInt;
-import java.util.Set;
-
-import javax.lang.model.element.AnnotationMirror;
 import javax.lang.model.element.Element;
 import javax.lang.model.element.PackageElement;
 import javax.lang.model.util.ElementFilter;
@@ -30,17 +22,25 @@ final class Utils {
    *
    * @param prism
    */
-  public static boolean isNonNullable(Element e, BuilderPrism prism) {
+  public static boolean isNonNullable(Element e) {
 
-    for (final AnnotationMirror mirror : UType.parse(e.asType()).annotations()) {
-      if (mirror.getAnnotationType().toString().endsWith("NonNull")) {
-        return true;
-      }
+    for (var mirror : UType.parse(e.asType()).annotations()) {
       if (mirror.getAnnotationType().toString().endsWith("Nullable")) {
         return false;
       }
     }
-    return prism.enforceNullSafety() || GlobalSettings.enforceNullSafety(e);
+
+    return checkNullMarked(e);
+  }
+
+  private static boolean checkNullMarked(Element e) {
+    var enclosing = e.getEnclosingElement();
+    if (enclosing == null || NullUnmarkedPrism.isPresent(enclosing)) {
+      return false;
+    } else if (NullMarkedPrism.isPresent(enclosing)) {
+      return true;
+    }
+    return checkNullMarked(enclosing);
   }
 
   public static boolean isNullableType(String type) {
