+++ avaje-record-builder-core/src/main/java/io/avaje/recordbuilder/internal/package-info.java
@@ -1,10 +1,10 @@
 @GeneratePrism(NullMarked.class)
 @GeneratePrism(NullUnmarked.class)
-@GeneratePrism(NonNullApi.class)
+@GeneratePrism(Nullable.class)
 package io.avaje.recordbuilder.internal;
 
 import org.jspecify.annotations.NullMarked;
 import org.jspecify.annotations.NullUnmarked;
+import org.jspecify.annotations.Nullable;
 
-import io.avaje.lang.NonNullApi;
 import io.avaje.prism.GeneratePrism;
