+++ avaje-record-builder-core/src/main/java/module-info.java
@@ -1,7 +1,6 @@
 module io.avaje.recordbuilder.core {
 
   requires java.compiler;
-  requires static io.avaje.lang;
   requires static io.avaje.prism;
   requires static org.jspecify;
