+++ avaje-record-builder/src/main/java/module-info.java
@@ -1,5 +1,5 @@
 module io.avaje.recordbuilder {
 
   exports io.avaje.recordbuilder;
-  requires transitive io.avaje.lang;
+  requires static transitive org.jspecify;
 }
