+++ blackbox-test-records/src/main/java/io/avaje/recordbuilder/defaults/Defaults.java
@@ -4,9 +4,8 @@ import java.util.List;
 
 import io.avaje.recordbuilder.DefaultValue;
 import io.avaje.recordbuilder.RecordBuilder;
-import io.avaje.recordbuilder.test.SomeInterface;
 
-@RecordBuilder(builderInterfaces = SomeInterface.class)
+@RecordBuilder
 public record Defaults(
     @DefaultValue("List.of(1,2,3)") List<Integer> list,
     @DefaultValue("24") int num,
