+++ blackbox-test-records/src/main/java/io/avaje/recordbuilder/nullable/Nullity.java
@@ -1,9 +1,8 @@
 package io.avaje.recordbuilder.nullable;
 
 import org.jspecify.annotations.NonNull;
-import org.jspecify.annotations.Nullable;
 
 import io.avaje.recordbuilder.RecordBuilder;
 
 @RecordBuilder
-public record Nullity(@NonNull String nonNull, @Nullable String nully, String marked) {}
+public record Nullity(@NonNull String nonNull, String nully, String marked) {}
